сделать для Comment отдельные папку, сервис, контроллер, использовать в CommentMapper @Builder
