+++ README.md
@@ -1,2 +1,2 @@
-# java-explore-with-me
-Template repository for ExploreWithMe project.
+Pull request for feature_comments
+https://github.com/nyxxyzu/java-explore-with-me/pull/3
