+++ main/src/main/java/ru/practicum/comment/CommentRepository.java
@@ -1,8 +1,7 @@
-package ru.practicum.event.repositories;
+package ru.practicum.comment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import ru.practicum.event.model.Comment;
 
 import org.springframework.data.domain.Pageable;
 import java.time.LocalDateTime;
