+++ main/src/main/java/ru/practicum/comment/dto/CommentDto.java
@@ -1,12 +1,14 @@
-package ru.practicum.event.dto.comment;
+package ru.practicum.comment.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Builder;
 import lombok.Data;
 import ru.practicum.user.dto.UserShortDto;
 
 import java.time.LocalDateTime;
 
 @Data
+@Builder
 public class CommentDto {
 
 	private Long id;
