+++ main/src/main/java/ru/practicum/event/EventService.java
@@ -1,8 +1,6 @@
 package ru.practicum.event;
 
 import ru.practicum.enums.Sort;
-import ru.practicum.event.dto.comment.CommentDto;
-import ru.practicum.event.dto.comment.NewCommentDto;
 import ru.practicum.event.dto.event.EventDto;
 import ru.practicum.event.dto.event.EventShortDto;
 import ru.practicum.event.dto.event.NewEventRequestDto;
@@ -44,19 +42,4 @@ public interface EventService {
 
 	EventRequestStatusUpdateResultDto changeRequestStatus(EventUpdateStatusUpdateRequestDto dto, Long userId, Long eventId);
 
-	CommentDto postComment(NewCommentDto dto, Long eventId, Long userId);
-
-	CommentDto editComment(NewCommentDto dto, Long userId, Long commentId);
-
-	void deleteCommentUser(Long userId, Long eventId, Long commentId);
-
-	List<CommentDto> getCommentsForUserEvent(Long userId, Long eventId);
-
-	void deleteCommentAdmin(Long eventId, Long commentId);
-
-	CommentDto getCommentById(Long commentId);
-
-	List<CommentDto> getCommentsByUser(Long userId);
-
-	List<CommentDto> getComments(LocalDateTime rangeStart, LocalDateTime rangeEnd, int from, int size);
 }
