+++ main/src/main/java/ru/practicum/event/EventServiceImpl.java
@@ -2,7 +2,6 @@ package ru.practicum.event;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.category.Category;
@@ -12,8 +11,6 @@ import ru.practicum.enums.RequestStatus;
 import ru.practicum.enums.Sort;
 import ru.practicum.enums.StateAction;
 import ru.practicum.enums.StateActionAdmin;
-import ru.practicum.event.dto.comment.CommentDto;
-import ru.practicum.event.dto.comment.NewCommentDto;
 import ru.practicum.event.dto.event.UpdateEventAdminRequestDto;
 import ru.practicum.event.dto.event.UpdateEventUserRequestDto;
 import ru.practicum.event.dto.request.ParticipationRequestDto;
@@ -24,11 +21,9 @@ import ru.practicum.event.dto.event.EventShortDto;
 import ru.practicum.event.dto.event.NewEventRequestDto;
 import ru.practicum.event.dto.request.EventRequestStatusUpdateResultDto;
 import ru.practicum.event.dto.request.EventUpdateStatusUpdateRequestDto;
-import ru.practicum.event.model.Comment;
 import ru.practicum.event.model.Event;
 import ru.practicum.event.model.Location;
 import ru.practicum.event.model.ParticipationRequest;
-import ru.practicum.event.repositories.CommentRepository;
 import ru.practicum.event.repositories.EventRepository;
 import ru.practicum.event.repositories.LocationRepository;
 import ru.practicum.event.repositories.RequestRepository;
@@ -39,7 +34,6 @@ import ru.practicum.user.User;
 import ru.practicum.user.UserRepository;
 import ru.practicum.StatsClient;
 
-import org.springframework.data.domain.Pageable;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -55,21 +49,18 @@ public class EventServiceImpl implements EventService {
 	private final LocationRepository locationRepository;
 	private final CategoryRepository categoryRepository;
 	private final RequestRepository requestRepository;
-	private final CommentRepository commentRepository;
 	private final StatsClient statsClient;
 
 	@Autowired
 	public EventServiceImpl(EventRepository eventRepository, UserRepository userRepository,
 							LocationRepository locationRepository, CategoryRepository categoryRepository,
-							RequestRepository requestRepository, StatsClient statsClient,
-							CommentRepository commentRepository) {
+							RequestRepository requestRepository, StatsClient statsClient) {
 		this.eventRepository = eventRepository;
 		this.userRepository = userRepository;
 		this.locationRepository = locationRepository;
 		this.categoryRepository = categoryRepository;
 		this.requestRepository = requestRepository;
 		this.statsClient = statsClient;
-		this.commentRepository = commentRepository;
 	}
 
 	@Override
@@ -390,105 +381,6 @@ public class EventServiceImpl implements EventService {
 		result.setConfirmedRequests(confirmedRequests);
 		result.setRejectedRequests(rejectedRequests);
 		return result;
-	}
-
-	@Override
-	@Transactional
-	public CommentDto postComment(NewCommentDto dto, Long userId, Long eventId) {
-		Event event = eventRepository.findById(eventId)
-				.orElseThrow(() -> new NotFoundException("Event not found"));
-		User user = userRepository.findById(userId)
-				.orElseThrow(() -> new NotFoundException("User not found"));
-		if (!event.getState().equals(State.PUBLISHED)) {
-			throw new ConditionsNotMetException("Cannot leave a comment on an unpublished event");
-		}
-		Comment comment = EventMapper.toComment(dto);
-		comment.setPosted(LocalDateTime.now());
-		comment.setPoster(user);
-		event.getComments().add(comment);
-		return EventMapper.toCommentDto(commentRepository.save(comment));
-	}
-
-	@Override
-	@Transactional
-	public CommentDto editComment(NewCommentDto dto, Long userId, Long commentId) {
-		if (!userRepository.existsById(userId)) {
-			throw new NotFoundException("User not found");
-		}
-		Comment comment = commentRepository.findById(commentId)
-				.orElseThrow(() -> new NotFoundException("Comment not found"));
-		if (!comment.getPoster().getId().equals(userId)) {
-			throw new ConditionsNotMetException("You can only edit your own comments");
-		}
-		comment.setText(dto.getText());
-		return EventMapper.toCommentDto(comment);
-	}
-
-	@Override
-	@Transactional
-	public void deleteCommentUser(Long userId, Long eventId, Long commentId) {
-		if (!userRepository.existsById(userId)) {
-			throw new NotFoundException("User not found");
-		}
-		Comment comment = commentRepository.findById(commentId)
-				.orElseThrow(() -> new NotFoundException("Comment not found"));
-		Event event = eventRepository.findById(eventId)
-				.orElseThrow(() -> new NotFoundException("Event not found"));
-		if (!comment.getPoster().getId().equals(userId)) {
-			throw new ConditionsNotMetException("You can only delete your own comments");
-		}
-		event.getComments().remove(comment);
-		commentRepository.deleteById(commentId);
-	}
-
-	@Override
-	public List<CommentDto> getCommentsForUserEvent(Long userId, Long eventId) {
-		if (!userRepository.existsById(userId)) {
-			throw new NotFoundException("User not found");
-		}
-		if (!eventRepository.existsById(eventId)) {
-			throw new NotFoundException("Event not found");
-		}
-		List<Comment> comments = commentRepository.findByUserIdAndEventId(userId, eventId);
-		return comments.stream()
-				.map(EventMapper::toCommentDto)
-				.toList();
-	}
-
-	@Override
-	@Transactional
-	public void deleteCommentAdmin(Long eventId, Long commentId) {
-		Comment comment = commentRepository.findById(commentId)
-				.orElseThrow(() -> new NotFoundException("Comment not found"));
-		Event event = eventRepository.findById(eventId)
-				.orElseThrow(() -> new NotFoundException("Event not found"));
-		event.getComments().remove(comment);
-		commentRepository.deleteById(commentId);
-	}
-
-	@Override
-	public CommentDto getCommentById(Long commentId) {
-		Comment comment = commentRepository.findById(commentId)
-				.orElseThrow(() -> new NotFoundException("Comment not found"));
-		return EventMapper.toCommentDto(comment);
-	}
-
-	@Override
-	public List<CommentDto> getCommentsByUser(Long userId) {
-		if (!userRepository.existsById(userId)) {
-			throw new NotFoundException("User not found");
-		}
-		return commentRepository.findByPosterId(userId).stream()
-				.map(EventMapper::toCommentDto)
-				.toList();
-	}
-
-	@Override
-	public List<CommentDto> getComments(LocalDateTime rangeStart, LocalDateTime rangeEnd, int from, int size) {
-		Pageable page = PageRequest.of(from, size);
-		return commentRepository.getComments(rangeStart, rangeEnd, page).stream()
-				.map(EventMapper::toCommentDto)
-				.toList();
 	}
 
 	private Long getViews(LocalDateTime start, LocalDateTime end, String uri) {
