+++ main/src/main/java/ru/practicum/event/controllers/EventControllerAdmin.java
@@ -4,7 +4,6 @@ import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import ru.practicum.event.EventServiceImpl;
-import ru.practicum.event.dto.comment.CommentDto;
+
 import ru.practicum.event.dto.event.UpdateEventAdminRequestDto;
 import ru.practicum.event.dto.event.EventDto;
 
@@ -52,34 +51,4 @@ public class EventControllerAdmin {
 		log.info("Updated event by id of " + eventId + " with data from " + request.toString());
 		return updatedEvent;
 	}
-
-	@DeleteMapping("/{eventId}/comments/{commentId}")
-	public void deleteCommentAdmin(@PathVariable("eventId") Long eventId,
-								   @PathVariable("commentId") Long commentId) {
-		log.info("Deleted comment {}, admin endpoint", commentId);
-		eventService.deleteCommentAdmin(eventId, commentId);
-	}
-
-	@GetMapping("/comments/{commentId}")
-	public CommentDto getCommentById(@PathVariable("commentId") Long commentId) {
-		return eventService.getCommentById(commentId);
-	}
-
-	@GetMapping("/comments")
-	public List<CommentDto> getComments(@DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") @RequestParam(required = false) LocalDateTime rangeStart,
-										@DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") @RequestParam(required = false) LocalDateTime rangeEnd,
-										@RequestParam(defaultValue = "0") int from,
-										@RequestParam(defaultValue = "10") int size) {
-		return eventService.getComments(rangeStart, rangeEnd, from, size);
-	}
-
-	@GetMapping("/comments/users/{userId}")
-	public List<CommentDto> getCommentsByUser(@PathVariable("userId") Long userId) {
-		return eventService.getCommentsByUser(userId);
-	}
-
-
-
-
-
 }
